requirements.txt
numpy
dv
argparse
matplotlib
opencv-python
tqdm
torch
torchvision
torch
PyYAML
pillow
scikit-learn
segmentation-models-pytorch
pycocotools
pandas